+++ src/components/logger/history.ts
@@ -30,6 +30,7 @@ import { ProbeRequestResponse } from '../../interfaces/request'
 import { Probe } from '../../interfaces/probe'
 import { Notification } from '../../interfaces/notification'
 import { log } from '../../utils/pino'
+import { getConfig } from '../config'
 const sqlite3 = SQLite3.verbose()
 const dbPath = path.resolve(process.cwd(), 'monika-logs.db')
 
@@ -189,7 +190,7 @@ export async function getUnreportedLogsCount(): Promise<number> {
   return row?.count || 0
 }
 
-export async function getUnreportedLogs(limit: number): Promise<UnreportedLog> {
+export async function getUnreportedLogs(ids: string[]): Promise<UnreportedLog> {
   const readUnreportedRequestsSQL = `
     SELECT PR.id,
       PR.created_at as timestamp,
@@ -209,9 +210,8 @@ export async function getUnreportedLogs(limit: number): Promise<UnreportedLog> {
       END alerts
     FROM probe_requests PR
       LEFT JOIN alerts A ON PR.id = A.probe_request_id
-    WHERE PR.reported = 0
-    GROUP BY PR.id
-    LIMIT ${limit};`
+    WHERE PR.reported = 0 AND PR.probe_id IN ('${ids.join("','")}')
+    GROUP BY PR.id;`
 
   const readUnreportedNotificationsSQL = `
     SELECT id,
@@ -223,8 +223,7 @@ export async function getUnreportedLogs(limit: number): Promise<UnreportedLog> {
       notification_id,
       channel
     FROM notifications
-    WHERE reported = 0
-    LIMIT ${limit};`
+    WHERE reported = 0 AND probe_id IN ('${ids.join("','")}');`
 
   const [unreportedRequests, unreportedNotifications] = await Promise.all([
     db.all(readUnreportedRequestsSQL).then(
@@ -248,32 +247,18 @@ export async function getUnreportedLogs(limit: number): Promise<UnreportedLog> {
   }
 }
 
-export async function setRequestLogAsReported(ids: number[]) {
-  const updateRowsSQL = `UPDATE probe_requests SET reported = 1 WHERE id IN (${ids.join(
-    ', '
-  )})`
+export async function deleteRequestLogs(ids: string[]) {
+  const idsString = ids.join("','")
+  const sql = `DELETE FROM probe_requests WHERE probe_id IN ('${idsString}');`
 
-  await db.run(updateRowsSQL)
+  return db.run(sql)
 }
 
-export async function deleteRequestLogs(ids: number[]) {
-  const sql = `DELETE FROM probe_requests WHERE id IN (${ids.join(', ')})`
+export async function deleteNotificationLogs(ids: string[]) {
+  const idsString = ids.join("','")
+  const sql = `DELETE FROM notifications WHERE probe_id IN ('${idsString}');`
 
-  await db.run(sql)
-}
-
-export async function setNotificationLogAsReported(ids: number[]) {
-  const updateRowsSQL = `UPDATE notifications SET reported = 1 WHERE id IN (${ids.join(
-    ', '
-  )})`
-
-  await db.run(updateRowsSQL)
-}
-
-export async function deleteNotificationLogs(ids: number[]) {
-  const sql = `DELETE FROM notifications WHERE id IN (${ids.join(', ')})`
-
-  await db.run(sql)
+  return db.run(sql)
 }
 
 /**
@@ -426,12 +411,10 @@ export async function saveNotificationLog(
 
 export async function getSummary() {
   const getNotificationsSummaryByTypeSQL = `SELECT type, COUNT(*) as count FROM notifications WHERE created_at > strftime('%s', datetime('now', '-24 hours')) GROUP BY type;`
-  const getProbesSummarySQL = `SELECT probe_id, COUNT(*) as count, AVG(response_time) as average_response_time FROM probe_requests WHERE created_at > strftime('%s', datetime('now', '-24 hours')) GROUP BY probe_id;`
 
-  const [notificationsSummaryByType, probesSummary] = await Promise.all([
-    db.all(getNotificationsSummaryByTypeSQL),
-    db.all(getProbesSummarySQL),
-  ])
+  const notificationsSummaryByType = await db.all(
+    getNotificationsSummaryByTypeSQL
+  )
 
   const numberOfIncidents: number =
     notificationsSummaryByType.find((notif) => notif.type === 'NOTIFY-INCIDENT')
@@ -444,8 +427,10 @@ export async function getSummary() {
     0
   )
 
+  const config = getConfig()
+
   return {
-    numberOfProbes: probesSummary.length,
+    numberOfProbes: config.probes.length,
     numberOfIncidents,
     numberOfRecoveries,
     numberOfSentNotifications,